+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 
 - Click outside dialog to dismiss it.
 - Fixed #603 Machine translation of tutorials.
+- Include all draft locales, but mark as draft.
 
 ### Fixed